Added cast to float32 in framestack.
FrameStack should cast to float32, otherwise there's an warning from
gymnasium about precision.

+++ baselines/common/atari_wrappers.py
@@ -199,7 +199,7 @@ class FrameStack(gym.Wrapper):
         self.k = k
         self.frames = deque([], maxlen=k)
         shp = env.observation_space.shape
-        self.observation_space = spaces.Box(low=0, high=255, shape=(shp[:-1] + (shp[-1] * k,)), dtype=env.observation_space.dtype)
+        self.observation_space = spaces.Box(low=np.float32(0), high=np.float32(255), shape=(shp[:-1] + (shp[-1] * k,)), dtype=env.observation_space.dtype)
 
     def reset(self, **kwargs):
         ob, res_info = self.env.reset(**kwargs)
